+++ conservatories/conservatories_model.go
@@ -1,7 +1,7 @@
 package conservatories
 
 type conservatory struct {
-	ID         int64   `json:"id"`
+	ID         uint    `json:"id"`
 	Name       string  `json:"name"`
 	Category   string  `json:"category"`
 	Street     string  `json:"street"`
@@ -12,7 +12,7 @@ type conservatory struct {
 	Longitude  float64 `json:"lng"`
 }
 
-func findConservatories(page, pageSize int64, sortKey string, ascending bool, query string) (int64, []*conservatory, error) {
+func findConservatories(page, pageSize uint, sortKey string, ascending bool, query string) (uint, []*conservatory, error) {
 	count, err := countConservatories(query)
 	if err != nil {
 		return 0, nil, err
